+++ src/components/DownloadButton/sub-buttons.tsx
@@ -22,7 +22,7 @@ export const DownloadSubButtons: React.FC<Props> = (props) => {
       <div className='p-6 grid grid-cols-2 gap-4 diviide-y-2 divide-white/[.1] text-left'>
         <button
           disabled={props.disabled}
-          className='inline-flex flex-col justify-center items-center bg-white/[.1] hover:bg-orange-400 fill-white/[.5] hover:fill-black/[.6] text-white/[.7] hover:text-black/[.6] rounded-xl p-4 text-center'
+          className='inline-flex flex-col justify-center items-center bg-white/[.1] hover:bg-orange-400 fill-white/[.5] hover:fill-black/[.6] text-white/[.7] hover:text-black/[.6] rounded-xl p-4 text-center transition active:scale-90'
           onClick={() => props.onClick('x11')}>
           <LinuxDownloadSVG />
           <p className='text-sm font-bold mt-2'>XCursors</p>
@@ -30,7 +30,7 @@ export const DownloadSubButtons: React.FC<Props> = (props) => {
         </button>
         <button
           disabled={props.disabled}
-          className='inline-flex flex-col justify-center items-center bg-white/[.1] hover:bg-blue-400 fill-white/[.5] hover:fill-black/[.6] text-white/[.7] hover:text-black/[.6] rounded-xl p-4 text-center'
+          className='inline-flex flex-col justify-center items-center bg-white/[.1] hover:bg-blue-400 fill-white/[.5] hover:fill-black/[.6] text-white/[.7] hover:text-black/[.6] rounded-xl p-4 text-center transition active:scale-90'
           onClick={() => props.onClick('win')}>
           <WindowsDownloadSVG />
           <p className='text-sm font-bold mt-2'>Windows Cursors</p>
@@ -39,7 +39,7 @@ export const DownloadSubButtons: React.FC<Props> = (props) => {
 
         <button
           disabled={props.disabled}
-          className='inline-flex col-span-2 flex-col justify-center items-center bg-white/[.1] hover:bg-violet-400 fill-white/[.5] hover:fill-black/[.6] text-white/[.7] hover:text-black/[.6] rounded-xl p-4 text-center'
+          className='inline-flex col-span-2 flex-col justify-center items-center bg-white/[.1] hover:bg-violet-400 fill-white/[.5] hover:fill-black/[.6] text-white/[.7] hover:text-black/[.6] rounded-xl p-4 text-center transition active:scale-90'
           onClick={() => props.onClick('png')}>
           <PNGsDownloadSVG />
           <p className='text-sm font-bold mt-2'>PNGs</p>