+++ src/components/TypePicker.tsx
@@ -22,10 +22,10 @@ export const TypePicker: React.FC<Props> = (props) => {
               title={`Bibata ${t}`}
               disabled={t === props.value}
               onClick={() => props.onClick(t)}
-              className={`py-2 flex justify-center items-center gap-1 font-bold border rounded-full shadow-md border-white/[.2] transition ${
+              className={`py-2 flex justify-center items-center gap-1 font-bold border rounded-full shadow-md border-white/[.2] transition active:scale-90 ${
                 t === props.value
                   ? 'bg-white fill-black text-black font-bold'
-                  : 'fill-white/[.4] text-white/[.4] hover:text-white hover:fill-white hover:border-white '
+                  : 'fill-white/[.4] text-white/[.4] hover:text-white hover:fill-white hover:border-white'
               }`}>
               <>
                 {t === 'Modern' && <ModernSVG />}